[Misc] Fixed a warning, updated formatting, replaced `_DEBUG` macros, started with ChangeLog-v0.05.md document.

+++ sources/Renderer/Direct3D12/RenderState/D3D12GraphicsPSO.cpp
@@ -304,7 +304,7 @@ void D3D12GraphicsPSO::CreateNativePSO(
     if (!inputElements_.empty())
     {
         stateDesc.InputLayout.pInputElementDescs = inputElements_.data();
-        stateDesc.InputLayout.NumElements        = inputElements_.size();
+        stateDesc.InputLayout.NumElements        = static_cast<UINT>(inputElements_.size());
     }
     else
     {
